[hotfix] #188 Modify refresh feature on the my tab
