+++ app/src/main/java/com/afume/afume_android/util/FragmentExtension.kt
@@ -1,0 +1,11 @@
+package com.afume.afume_android.util
+
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentTransaction
+
+fun Fragment.refreshFragment(fragment: Fragment, fragmentManager: FragmentManager) {
+    var ft: FragmentTransaction = fragmentManager.beginTransaction()
+
+    ft.detach(fragment).attach(fragment).commit()
+}
